refactor: also exporting types
